+++ routes/events.go
@@ -1,0 +1,69 @@
+package routes
+
+import (
+	"net/http"
+	"strconv"
+	"udemy-rest-api/db"
+	"udemy-rest-api/models"
+
+	"github.com/gin-gonic/gin"
+)
+
+func getEvents(context *gin.Context) {
+	events, err := models.GetAllEvents()
+	if err != nil {
+		context.JSON(http.StatusInternalServerError, gin.H{"message": "Could not fetch events"})
+		return
+	}
+	context.JSON(http.StatusOK, events)
+}
+
+func getEvent(context *gin.Context) {
+	id := context.Param("id")
+	eventId, err := strconv.ParseInt(id, 10, 64)
+	if err != nil {
+		context.JSON(http.StatusBadRequest, gin.H{"message": "Invalid event ID"})
+		return
+	}
+
+	event, err := models.GetEventById(eventId)
+	if err != nil {
+		context.JSON(http.StatusInternalServerError, gin.H{"message": "Could not fetch event"})
+		return
+	}
+
+	context.JSON(http.StatusOK, event)
+}
+
+func createEvent(context *gin.Context) {
+	var event models.Event
+	err := context.ShouldBindJSON(&event)
+
+	if err != nil {
+		context.JSON(http.StatusBadRequest, gin.H{"message": "Could not parse request body"})
+		return
+	}
+
+	event.ID = 1
+	event.UserID = 1
+	event.Save()
+
+	if err != nil {
+		context.JSON(http.StatusInternalServerError, gin.H{"message": "Could not create event"})
+		return
+	}
+
+	context.JSON(http.StatusCreated, gin.H{"message": "Event created successfully", "event": event})
+}
+
+func GetEventById(id int64) *models.Event {
+	query := "SELECT * FROM events WHERE id = ?"
+	row := db.DB.QueryRow(query, id)
+
+	var event models.Event
+	err := row.Scan(&event.ID, &event.Name, &event.Description, &event.Location, &event.DateTime, &event.UserID)
+	if err != nil {
+		return nil
+	}
+	return &event
+}
